+++ kamu-cli/src/cli_parser.rs
@@ -932,7 +932,7 @@ pub fn cli() -> App<'static, 'static> {
                 ))
             ),
             SubCommand::with_name("verify")
-                .about("Verifies the validity of derivative data")
+                .about("Verifies the validity of a dataset")
                 .args(&[
                     Arg::with_name("recursive")
                         .short("r")
@@ -956,6 +956,8 @@ pub fn cli() -> App<'static, 'static> {
                     This command can help you with the last stage. It uses the history of transformations stored in metadata to first compare the hashes of data with ones stored in metadata (i.e. verify that data corresponds to metadata). Then it repeats all declared transformations locally to ensure that what's declared in metadata actually produces the presented result.
 
                     The combination of the above steps can give you a high certainty that the data you're using is trustworthy.
+
+                    When called on a root dataset the command will only perform the integrity check of comparing data hashes to metadata.
 
                     ### Examples ###