+++ kamu-core/src/domain/transform_service.rs
@@ -7,11 +7,13 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
-use super::{DomainError, EngineError, EngineProvisioningListener};
-use opendatafabric::{DatasetID, DatasetIDBuf, MetadataBlock, Sha3_256};
+use super::{
+    EngineError, EngineProvisioningListener, VerificationError, VerificationListener,
+    VerificationMultiListener, VerificationOptions, VerificationRequest, VerificationResult,
+};
+use opendatafabric::{DatasetID, DatasetIDBuf, Sha3_256};
 
 use std::backtrace::Backtrace;
-use std::fmt::Display;
 use std::sync::Arc;
 use std::usize;
 use thiserror::Error;
@@ -33,7 +35,7 @@ pub trait TransformService: Send + Sync {
         listener: Option<Arc<dyn TransformMultiListener>>,
     ) -> Vec<(DatasetIDBuf, Result<TransformResult, TransformError>)>;
 
-    fn verify(
+    fn verify_transform(
         &self,
         dataset_id: &DatasetID,
         block_range: (Option<Sha3_256>, Option<Sha3_256>),
@@ -41,7 +43,7 @@ pub trait TransformService: Send + Sync {
         listener: Option<Arc<dyn VerificationListener>>,
     ) -> Result<VerificationResult, VerificationError>;
 
-    fn verify_multi(
+    fn verify_transform_multi(
         &self,
         datasets: &mut dyn Iterator<Item = VerificationRequest>,
         options: VerificationOptions,
@@ -61,32 +63,6 @@ pub enum TransformResult {
         new_head: Sha3_256,
         num_blocks: usize,
     },
-}
-
-///////////////////////////////////////////////////////////////////////////////
-
-#[derive(Debug)]
-pub struct VerificationRequest<'a> {
-    pub dataset_id: &'a DatasetID,
-    pub block_range: (Option<Sha3_256>, Option<Sha3_256>),
-}
-
-///////////////////////////////////////////////////////////////////////////////
-
-#[derive(Debug)]
-pub enum VerificationResult {
-    Valid { blocks_verified: usize },
-}
-
-///////////////////////////////////////////////////////////////////////////////
-
-#[derive(Debug, Clone)]
-pub struct VerificationOptions;
-
-impl Default for VerificationOptions {
-    fn default() -> Self {
-        Self
-    }
 }
 
 ///////////////////////////////////////////////////////////////////////////////
@@ -120,47 +96,6 @@ pub struct NullTransformMultiListener;
 impl TransformMultiListener for NullTransformMultiListener {}
 
 ///////////////////////////////////////////////////////////////////////////////
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum VerificationPhase {
-    HashData,
-    ReplayTransform,
-    BlockValid,
-}
-
-pub trait VerificationListener {
-    fn begin(&self, _num_blocks: usize) {}
-    fn on_phase(
-        &self,
-        _block: &Sha3_256,
-        _block_index: usize,
-        _num_blocks: usize,
-        _phase: VerificationPhase,
-    ) {
-    }
-    fn success(&self, _result: &VerificationResult) {}
-    fn error(&self, _error: &VerificationError) {}
-
-    fn get_transform_listener(self: Arc<Self>) -> Option<Arc<dyn TransformListener>> {
-        None
-    }
-}
-
-pub struct NullVerificationListener;
-impl VerificationListener for NullVerificationListener {}
-
-///////////////////////////////////////////////////////////////////////////////
-
-pub trait VerificationMultiListener {
-    fn begin_verify(&self, _dataset_id: &DatasetID) -> Option<Arc<dyn VerificationListener>> {
-        None
-    }
-}
-
-pub struct NullVerificationMultiListener;
-impl VerificationMultiListener for NullVerificationMultiListener {}
-
-///////////////////////////////////////////////////////////////////////////////
 // Errors
 ///////////////////////////////////////////////////////////////////////////////
 
@@ -184,60 +119,5 @@ impl TransformError {
             source: e.into(),
             backtrace: Backtrace::capture(),
         }
-    }
-}
-
-///////////////////////////////////////////////////////////////////////////////
-
-#[derive(Debug, Error)]
-pub enum VerificationError {
-    #[error("Dataset is not derivative")]
-    NotDerivative,
-    #[error("Block {0} not found")]
-    NoSuchBlock(Sha3_256),
-    #[error("Data doesn't match metadata: {0}")]
-    DataDoesNotMatchMetadata(DataDoesNotMatchMetadata),
-    #[error("Data is not reproducible: {0}")]
-    DataNotReproducible(DataNotReproducible),
-    #[error("Tranform error: {0}")]
-    TransformError(#[from] TransformError),
-    #[error("Domain error: {0}")]
-    DomainError(#[from] DomainError),
-}
-
-///////////////////////////////////////////////////////////////////////////////
-
-#[derive(Debug)]
-pub struct DataDoesNotMatchMetadata {
-    pub block_hash: Sha3_256,
-    pub expected_hash: String,
-    pub actual_hash: String,
-}
-
-impl Display for DataDoesNotMatchMetadata {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(
-            f,
-            "Data hash for block {} is expected to be {} but actual {}",
-            self.block_hash, self.expected_hash, self.actual_hash
-        )
-    }
-}
-
-///////////////////////////////////////////////////////////////////////////////
-
-#[derive(Debug)]
-pub struct DataNotReproducible {
-    pub expected_block: MetadataBlock,
-    pub actual_block: MetadataBlock,
-}
-
-impl Display for DataNotReproducible {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(
-            f,
-            "Expected block {:?} but got non-equivalent block {:?}",
-            self.expected_block, self.actual_block
-        )
     }
 }